rag_project/requirements.txt
sentence-transformers
faiss-cpu
pymupdf
python-docx
ollama
aiohttp
numpy
python-multipart
uvicorn
fastapi